Ouvrir Pinpoint par lien profond pinpoint:// (#58)
Un client léger — un CLI, demain un serveur MCP — ne peut pas capturer l'écran
lui-même : macOS accorde l'autorisation « Enregistrement de l'écran » au bundle
que l'utilisateur a autorisé nommément, pas à un exécutable nu. Il lui faut donc
un moyen de demander la capture à l'app déjà lancée, qui, elle, détient le
droit. C'est ce que ce scheme apporte.

`CFBundleURLTypes` est déclaré dans project.yml, pas dans Info.plist : ce
dernier est réécrit à chaque `xcodegen generate` (même piège que les clés de
version, déjà commenté au-dessus).

Trois URL, et rien d'autre :

  pinpoint://capture                  démarre la sélection de région (= ⌘⇧1)
  pinpoint://last                     rouvre le dernier transfert dans l'éditeur
  pinpoint://last?format=json|md|png  révèle ce fichier dans le Finder

Une URL peut venir de n'importe où — un script, un terminal, une page web
simplement visitée — et LaunchServices ne dit pas laquelle. La réponse tient
dans URLCommand : chaque URL nomme une action et ne transporte rien d'autre
(pas de coordonnées, pas de chemin de sortie), `capture` n'ouvre que l'overlay
qu'il faut dessiner à la main puis un éditeur où il faut appuyer sur Copier, et
rien ne répond jamais à l'appelant. Aucune URL ne prend de photo toute seule :
c'est pourquoi la capture plein écran du menu n'a délibérément pas de lien.
Tout le reste est ignoré en silence.

L'app est LSUIElement, donc aucune fenêtre n'est ouverte quand l'URL arrive :
l'overlay s'affiche au niveau `.screenSaver` et le contrôleur de sélection
active l'app, et l'éditeur fait de même. Une URL reçue avant la fin du lancement
(démarrage à froid) est mise en file et rejouée à la fin de
`applicationDidFinishLaunching`, sinon elle déclencherait une capture avant que
l'item de menu et les observateurs existent.

`pinpoint://last` réutilise la notification `pinpointOpenInEditor` de l'étagère,
avec une réserve : un fichier venu de `last/` ne se voit plus passer de
`sourceURL`. Ce dossier est remplacé en bloc à chaque copie, donc le
`-cropped.png` qu'un recadrage y déposerait pollue ce qu'un agent lit et
disparaît sans prévenir.

Le contrat de transfert est scindé en deux : `HandoffContract.swift`
(où vivent les fichiers, comment les relire) et `HandoffDocument.swift` (les
types), séparés de l'écriture et de la construction. Le gestionnaire d'URL a
besoin des chemins sans rien savoir du rendu — et c'est ce qui permettra au CLI
de #56 de compiler le même contrat plutôt que d'en recopier une version.

Closes #58
